// apps/kid-church/src/models/KidChurch.ts
export enum KidGuardianRelationCodeEnum {
  FATHER = 'FATHER', // Padre
  MOTHER = 'MOTHER', // Madre
  BROTHER = 'BROTHER', // Hermano
  SISTER = 'SISTER', // Hermana
  GRANDFATHER = 'GRANDFATHER', // Abuelo
  GRANDMOTHER = 'GRANDMOTHER', // Abuela
  UNCLE = 'UNCLE', // Tio
  AUNT = 'AUNT', // Tia
  MALE_COUSIN = 'MALE_COUSIN', // Primo
  FEMALE_COUSIN = 'FEMALE_COUSIN', // Prima
  STEPMOTHER = 'STEPMOTHER', // Madrastra
  STEPFATHER = 'STEPFATHER', // Padrastro
  MALE_ACQUAINTANCE = 'MALE_ACQUAINTANCE', // Conocido
  FEMALE_ACQUAINTANCE = 'FEMALE_ACQUAINTANCE', // Conocida
}

export enum KidGuardianRelationEnum {
  FATHER = 'Padre', // Padre
  MOTHER = 'Madre', // Madre
  BROTHER = 'Hermano', // Hermano
  SISTER = 'Hermana', // Hermana
  GRANDFATHER = 'Abuelo', // Abuelo
  GRANDMOTHER = 'Abuela', // Abuela
  UNCLE = 'Tío', // Tío
  AUNT = 'Tía', // Tía
  MALE_COUSIN = 'Primo', // Primo
  FEMALE_COUSIN = 'Prima', // Prima
  STEPMOTHER = 'Madrastra', // Madrastra
  STEPFATHER = 'Padrastro', // Padrastro
  MALE_ACQUAINTANCE = 'Conocido', // Conocido
  FEMALE_ACQUAINTANCE = 'Conocida', // Conocida
}
// ...
export const kidRelationSelect = [
  {
    value: KidGuardianRelationCodeEnum.FATHER,
    label: KidGuardianRelationEnum.FATHER,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.MOTHER,
    label: KidGuardianRelationEnum.MOTHER,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.BROTHER,
    label: KidGuardianRelationEnum.BROTHER,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.SISTER,
    label: KidGuardianRelationEnum.SISTER,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.GRANDFATHER,
    label: KidGuardianRelationEnum.GRANDFATHER,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.GRANDMOTHER,
    label: KidGuardianRelationEnum.GRANDMOTHER,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.UNCLE,
    label: KidGuardianRelationEnum.UNCLE,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.AUNT,
    label: KidGuardianRelationEnum.AUNT,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.MALE_COUSIN,
    label: KidGuardianRelationEnum.MALE_COUSIN,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.FEMALE_COUSIN,
    label: KidGuardianRelationEnum.FEMALE_COUSIN,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.STEPMOTHER,
    label: KidGuardianRelationEnum.STEPMOTHER,
    gender: UserGenderCode.FEMALE,
  },
  {
    value: KidGuardianRelationCodeEnum.STEPFATHER,
    label: KidGuardianRelationEnum.STEPFATHER,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.MALE_ACQUAINTANCE,
    label: KidGuardianRelationEnum.MALE_ACQUAINTANCE,
    gender: UserGenderCode.MALE,
  },
  {
    value: KidGuardianRelationCodeEnum.FEMALE_ACQUAINTANCE,
    label: KidGuardianRelationEnum.FEMALE_ACQUAINTANCE,
    gender: UserGenderCode.FEMALE,
  },
];